[Palette creation] Adapt the palette creation to the kid flavor.
In the kid version, the instruction is represented as an image, and a default name is used.
In the default version, the instruction is in plain text, and the name of the palette
is choose by the user from a dialog.
