+++ CameraColorPicker/app/src/adult/res/layout/activity_palette_creation.xml
