+++ CameraColorPicker/app/src/adult/res/values/strings.xml
@@ -6,5 +6,18 @@
 
     <!-- Delete palette dialog fragment -->
     <string name="fragment_dialog_delete_palette_title">Warning</string>
-    <string name="fragment_dialog_delete_palette_message">Delete the palette %1$s from your phone?</string>
+    <string name="fragment_dialog_delete_palette_message">Delete the palette %1$s from your phone?
+    </string>
+
+    <!-- Palette creation activity -->
+    <string name="activity_palette_creation_instruction">To add colors to your palette\nsimply click
+        on the list bellow.
+    </string>
+    <string name="activity_palette_creation_edit_text_dialog_fragment_title">Create your palette
+    </string>
+    <string name="activity_palette_creation_edit_text_dialog_fragment_hint">My new awesome palette
+    </string>
+    <string name="activity_palette_creation_edit_text_dialog_fragment_positive_button">Create
+    </string>
+
 </resources>
