+++ CameraColorPicker/app/src/kids/res/layout/activity_palette_creation.xml
@@ -1,0 +1,62 @@
+<?xml version="1.0" encoding="utf-8"?>
+<merge xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <ImageView
+        android:id="@+id/activity_palette_creation_instruction"
+        android:layout_width="match_parent"
+        android:layout_height="150dp"
+        android:alpha="0.3"
+        android:padding="@dimen/default_padding"
+        android:scaleType="centerInside"
+        android:src="@drawable/drop_sad" />
+
+    <fr.tvbarthel.apps.cameracolorpicker.views.PaletteMakerView
+        android:id="@+id/activity_palette_creation_color_palette_builder"
+        android:layout_width="match_parent"
+        android:layout_height="150dp" />
+
+    <FrameLayout
+        android:layout_width="match_parent"
+        android:layout_height="150dp"
+        android:background="@android:color/transparent">
+
+        <ImageButton
+            android:id="@+id/activity_palette_creation_remove_button"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="end|top"
+            android:background="?attr/selectableItemBackgroundBorderless"
+            android:padding="8dp"
+            android:src="@drawable/ic_action_delete" />
+
+    </FrameLayout>
+
+    <View
+        android:layout_width="match_parent"
+        android:layout_height="8dp"
+        android:layout_marginTop="150dp"
+        android:background="@drawable/linear_gradient_bottom" />
+
+    <android.support.v7.widget.RecyclerView
+        android:id="@+id/activity_palette_creation_list_view"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_marginTop="150dp"
+        tools:listitem="@layout/row_color_item" />
+
+    <com.melnykov.fab.FloatingActionButton
+        android:id="@+id/activity_palette_creation_fab"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="bottom|end"
+        android:layout_margin="16dp"
+        android:src="@drawable/ic_check_white"
+        app:fab_colorNormal="@color/color_accent"
+        app:fab_colorPressed="@color/color_accent"
+        app:fab_colorRipple="?attr/colorControlHighlight" />
+
+</merge>
