+++ CameraColorPicker/app/src/kids/res/values-fr/strings.xml
@@ -2,4 +2,7 @@
 <resources>
     <string name="app_name">Kids Camera Color Picker</string>
     <string name="activity_main_error_not_enough_colors"></string>
+
+    <!-- Palette creation activity -->
+    <string name="activity_palette_creation_default_palette_name">Palette %d</string>
 </resources>
