+++ CameraColorPicker/app/src/main/res/values-fr/strings.xml
@@ -25,11 +25,6 @@
 
     <!-- Palette creation activity -->
     <string name="activity_palette_creation_label">Nouvelle palette</string>
-    <string name="activity_palette_creation_instruction">Pour ajouter des couleurs à votre palette\ncliquez sur la liste en dessus.</string>
-    <string name="activity_palette_creation_empty_color_palette_builder">Ajoutez au moins 1 couleur à la palette.</string>
-    <string name="activity_palette_creation_edit_text_dialog_fragment_title">Créez votre palette</string>
-    <string name="activity_palette_creation_edit_text_dialog_fragment_hint">Ma nouvelle palette</string>
-    <string name="activity_palette_creation_edit_text_dialog_fragment_positive_button">Créer</string>
 
     <!-- Palette detail activity -->
     <string name="activity_palette_detail_label">Detail</string>
