+++ CameraColorPicker/app/src/main/res/values/strings.xml
@@ -27,11 +27,6 @@
 
     <!-- Palette creation activity -->
     <string name="activity_palette_creation_label">New palette</string>
-    <string name="activity_palette_creation_instruction">To add colors to your palette\nsimply click on the list bellow.</string>
-    <string name="activity_palette_creation_empty_color_palette_builder">Add at least 1 color to the palette.</string>
-    <string name="activity_palette_creation_edit_text_dialog_fragment_title">Create your palette</string>
-    <string name="activity_palette_creation_edit_text_dialog_fragment_hint">My new awesome palette</string>
-    <string name="activity_palette_creation_edit_text_dialog_fragment_positive_button">Create</string>
 
     <!-- Palette detail activity -->
     <string name="activity_palette_detail_label">Detail</string>
